# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MoliCasillas",
        "description": "Documentación de la API para el Sistema de Casilla Electrónica MoliCasillas",
        "contact": {
            "email": "admin@molicasillas.gob.pe"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://molicasillas.test",
            "description": "Servidor de API MoliCasillas"
        }
    ],
    "paths": {
...
        "/api/v1/casilla/messages": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Listar mensajes",
                "operationId": "579e6cf706618ab0197eadd3f1a52b5f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "asunto",
                        "in": "query",
                        "description": "Filtrar por asunto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "description": "Filtrar por fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "description": "Filtrar por fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "etiqueta_estado",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SIN LEER",
                                "NOTIFICADO",
                                "LEÍDO",
                                "ARCHIVADO"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de mensajes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Mensaje"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/casilla/messages/statuses": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Contar mensajes por estado",
                "operationId": "c7aff56cd1acec9ea8b638f9814b30d2",
                "responses": {
                    "200": {
                        "description": "Conteo de estados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "estados": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "codigo": {
                                                                "type": "string"
                                                            },
                                                            "etiqueta": {
                                                                "type": "string"
                                                            },
                                                            "cantidad": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "cantidad_no_leidos": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/casilla/messages/{id}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener detalle del mensaje",
                "operationId": "2e5e59be0a362b688f27c3a81ed58a85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del mensaje",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mensaje"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/casilla/messages/{id}/read": {
            "patch": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Marcar mensaje como leído",
                "operationId": "3fce4e4aa077ad66548d4ae77c4e3be2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje marcado como leído",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mensaje"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/casilla/messages/{id}/archive": {
            "patch": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Archivar mensaje",
                "operationId": "3171bc595a25213bcec33de65725f533",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje archivado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mensaje"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Mensaje": {
                "title": "Mensaje",
                "description": "Mensaje de usuario",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "casilla_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "remitente_nombre": {
                        "type": "string"
                    },
                    "remitente_entidad": {
                        "type": "string"
                    },
                    "destinatario_nombre": {
                        "type": "string"
                    },
                    "destinatario_tipo_doc": {
                        "type": "string"
                    },
                    "destinatario_num_doc": {
                        "type": "string"
                    },
                    "asunto": {
                        "type": "string"
                    },
                    "cuerpo": {
                        "type": "string"
                    },
                    "registrado_en": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "codigo_estado": {
                        "type": "string"
                    },
                    "etiqueta_estado": {
                        "type": "string"
                    },
                    "notificado_en": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "leido_en": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "archivado_en": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "codigo_referencia": {
                        "type": "string",
                        "nullable": true
                    },
                    "codigo_expediente": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}